# docs/swagger.yaml
definitions:
  controller.GetAccountResp:
    $ref: '#/definitions/model.Account'
  controller.LoginReq:
    properties:
      email:
        type: string
      pwd:
        type: string
    required:
    - email
    - pwd
    type: object
  controller.LogoutReq:
    properties:
      cid:
        type: integer
    required:
    - cid
    type: object
  controller.TransferReq:
    properties:
      from:
        description: 转出方account_id
        type: string
      fromUnit:
        description: 转出方货币单位
        type: integer
      requestId:
        description: 调用方自定义请求ID，可以为空
        type: string
      to:
        description: 转入方account_id
        type: string
      toNum:
        description: 需要转入的货币数量，以转入方货币单位计算
        type: number
      toUnit:
        description: 转入方货币单位
        type: integer
    required:
    - from
    - fromUnit
    - to
    - toNum
    - toUnit
    type: object
  controller.TransferResp:
    properties:
      cost:
        description: 转账消耗金额，货币单位以请求中FromUnit计算
        type: number
      from:
        description: 转出方account_id
        type: string
      fromUnit:
        description: 转出方货币单位
        type: integer
      rate:
        description: 货币换算汇率 1转出方单位货币 = x转入方单位货币
        type: number
      requestId:
        description: 调用方自定义请求ID，可以为空
        type: string
      tid:
        description: 转账ID
        type: string
      to:
        description: 转入方account_id
        type: string
      toNum:
        description: 需要转入的货币数量，以转入方货币单位计算
        type: number
      toUnit:
        description: 转入方货币单位
        type: integer
    required:
    - cost
    - from
    - fromUnit
    - rate
    - tid
    - to
    - toNum
    - toUnit
    type: object
  model.Account:
    properties:
      accountId:
        description: 账户ID
        type: string
      balance:
        description: 账户余额
        type: number
      createdAt:
        description: 创建时间
        type: string
      modifiedAt:
        description: 创建时间
        type: string
      uid:
        description: 账户属主ID，同User.ID相同
        type: integer
      unit:
        description: 账户余额货币单位
        type: integer
    type: object
  restapi.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /v1/clients/{cid}/account/{aid}:
    get:
      description: 获取指定账户信息
      parameters:
      - description: 账户
        in: path
        name: aid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetAccountResp'
        "401":
          description: 授权失败/未授权
          schema:
            $ref: '#/definitions/restapi.Response'
        "404":
          description: 账户不存在
          schema:
            $ref: '#/definitions/restapi.Response'
        "406":
          description: 参数错误
          schema:
            $ref: '#/definitions/restapi.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/restapi.Response'
      summary: 获取指定账户信息
  /v1/clients/{cid}/account/{aid}/transfer:
    post:
      description: 用户指定账户向指定账户转账
      parameters:
      - description: 转账请求信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.TransferReq'
      - description: 用户ID
        in: path
        name: cid
        required: true
        type: integer
      - description: 用户账户ID
        in: path
        name: aid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TransferResp'
        "404":
          description: 账户不存在
          schema:
            $ref: '#/definitions/restapi.Response'
        "406":
          description: 参数错误
          schema:
            $ref: '#/definitions/restapi.Response'
        "451":
          description: 账户余额不足
          schema:
            $ref: '#/definitions/restapi.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/restapi.Response'
      summary: 转账
  /v1/login:
    post:
      description: 用户登录
      parameters:
      - description: 账户及密码
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restapi.Response'
        "401":
          description: 授权失败/未授权
          schema:
            $ref: '#/definitions/restapi.Response'
        "404":
          description: 账户不存在
          schema:
            $ref: '#/definitions/restapi.Response'
        "406":
          description: 参数错误
          schema:
            $ref: '#/definitions/restapi.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/restapi.Response'
      summary: 登录
  /v1/logout:
    post:
      description: 用户登出
      parameters:
      - description: 账户ID
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.LogoutReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restapi.Response'
        "401":
          description: 授权失败/未授权
          schema:
            $ref: '#/definitions/restapi.Response'
        "404":
          description: 账户不存在
          schema:
            $ref: '#/definitions/restapi.Response'
        "406":
          description: 参数错误
          schema:
            $ref: '#/definitions/restapi.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/restapi.Response'
      summary: 登出
swagger: "2.0"